+++ core/__init__.py
@@ -1,3 +1,3 @@
 # pyright: strict
-__version__ = "1.0.0rc7"
+__version__ = "1.0.0rc8"
 __framework__ = "Forge"