+++ packages/forge-mvc-entities/forge_mvc_entities/__init__.py
@@ -72,4 +72,4 @@ __all__ = [
     "to_snake",
 ]
 
-__version__ = "1.0.0rc7"
+__version__ = "1.0.0rc8"