+++ packages/forge-mvc-fixtures/forge_mvc_fixtures/__init__.py
@@ -24,6 +24,6 @@ from forge_mvc_fixtures.factory import (
     FixtureReference,
 )
 
-__version__ = "1.0.0rc7"
+__version__ = "1.0.0rc8"
 
 __all__ = ["Factory", "FactoryError", "Fixture", "FixtureReference"]