+++ packages/forge-mvc-i18n/forge_mvc_i18n/__init__.py
@@ -43,7 +43,7 @@ from forge_mvc_i18n.translator import (
     trans,
 )
 
-__version__ = "1.0.0rc7"
+__version__ = "1.0.0rc8"
 
 __all__ = [
     "I18nError",