+++ packages/forge-mvc-jobs/forge_mvc_jobs/__init__.py
@@ -34,7 +34,7 @@ from forge_mvc_jobs.queue import (
     run_worker,
 )
 
-__version__ = "1.0.0rc7"
+__version__ = "1.0.0rc8"
 
 __all__ = [
     "JobError",