+++ packages/forge-mvc-mail/forge_mvc_mail/__init__.py
@@ -60,4 +60,4 @@ __all__ = [
     "TransportResult",
 ]
 
-__version__ = "1.0.0rc7"
+__version__ = "1.0.0rc8"