+++ packages/forge-mvc-sessions-db/forge_mvc_sessions_db/__init__.py
@@ -43,4 +43,4 @@ __all__ = [
     "SessionMetrics",
 ]
 
-__version__ = "1.0.0rc7"
+__version__ = "1.0.0rc8"